requirements.txt
langchain>=0.1.0
langchain-google-genai>=0.0.6
chromadb>=0.4.22
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
pypdf2>=3.0.1
numpy>=1.24.3
pandas>=2.0.3
google-generativeai>=0.3.0
streamlit>=1.28.0
